feat(frontend): Add iCal Export modal
