+++ apps/frontend/src/app/poll/ical/ical.component.html
@@ -1,0 +1,13 @@
+<ngbx-modal #modal [back]="['..']">
+  <ng-container modal-title>
+    iCal Export
+  </ng-container>
+  <ng-container modal-body>
+    <p>Export your calendar to iCal format.</p>
+    <p>Click the button below to download the file.</p>
+  </ng-container>
+  <ng-container modal-footer>
+    <button class="btn btn-secondary" (click)="modal.close()">Close</button>
+    <button class="btn btn-primary" (click)="export()">Export</button>
+  </ng-container>
+</ngbx-modal>
